only run bank for mysql, ignore kafka.

+++ tests/bank/run.sh
@@ -9,6 +9,10 @@ CDC_BINARY=cdc.test
 SINK_TYPE=$1
 
 function prepare() {
+    if [ "$SINK_TYPE" != "kafka" ]; then
+      return
+    fi
+
     rm -rf $WORK_DIR && mkdir -p $WORK_DIR
 
     start_tidb_cluster --workdir $WORK_DIR
